requirements.txt
opencv-python
mediapipe
numpy
scipy
pyttsx3
pyyaml